+++ _import/tycho-execution/foundry/scripts/deploy-executors.js
@@ -99,6 +99,18 @@ const executors_to_deploy = {
                 "0x000000000022D473030F116dDEE9F6B43aC78BA3"
             ]
         },
+        // Args: permit2
+        {
+            exchange: "ERC4626Executor", args: [
+                "0x000000000022D473030F116dDEE9F6B43aC78BA3"
+            ]
+        },
+        // Args: permit2
+        {
+            exchange: "RocketpoolExecutor", args: [
+                "0x000000000022D473030F116dDEE9F6B43aC78BA3"
+            ]
+        },
     ],
     "base": [
         // Args: Factory, Pool Init Code Hash, Permit2, Fee BPS